maded progress dialog uncloseable, changed IPhase phsnum from int to long

+++ SysconCommon/Accounting/IPhase.cs
@@ -8,6 +8,6 @@ namespace SysconCommon.Accounting
     public interface IPhase
     {
         IJob Job { get; }
-        int PhaseNumber { get; }
+        long PhaseNumber { get; }
     }
 }
